# apps/web/src/lib/service-clients/service-properties/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "properties_service",
    "description": "",
    "termsOfService": "https://macro.com/terms",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/properties/tags": {
      "get": {
        "tags": ["Properties"],
        "summary": "List the caller's tag sets: their personal set, plus their team's set when on a team.\nPure read - a scope with no provisioned definition yet returns an empty set.",
        "operationId": "list_tags",
        "responses": {
          "200": {
            "description": "Tag sets retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagSetResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Properties"],
        "summary": "Provision (get-or-create) the caller's tag set for a scope and return it. Called when\nthe caller creates their first label for that scope, so the read path stays side-effect free.",
        "operationId": "ensure_tag_set",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnsureTagSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tag set provisioned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagSetResponse"
                }
              }
            }
          },
          "403": {
            "description": "Team membership required for team scope"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/tags/merge": {
      "post": {
        "tags": ["Properties"],
        "summary": "Replace one of the caller's personal labels with an existing team label.",
        "description": "Every entity carrying the personal label is retagged with the team label\n(deduped if it already has both) and the personal label is deleted. The team\nlabel's name and color win. This is the confirmation step after the promote\nendpoint reports a name collision.",
        "operationId": "merge_tag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Label merged into the team label",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Team membership required"
          },
          "404": {
            "description": "Either label was not found in the expected tag set"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/tags/promote": {
      "post": {
        "tags": ["Properties"],
        "summary": "Share one of the caller's personal labels with their team.",
        "description": "The label moves into the team tag set keeping its option id, so every entity\nalready carrying it keeps the label — it just becomes visible to, and usable\nby, the whole team. A 409 means the team already has a label with that name;\nits body carries that label, and confirming the replacement is a call to the\nmerge endpoint with `conflicting_option.id` as `target_option_id`.",
        "operationId": "promote_tag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Label moved to the team tag set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOptionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Team membership required"
          },
          "404": {
            "description": "Label not found in the caller's tag set"
          },
          "409": {
            "description": "The team already has a label with that name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagPromotionConflictResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DataType": {
        "type": "string",
        "description": "Data type for property values, determining storage and validation.",
        "enum": [
          "BOOLEAN",
          "DATE",
          "NUMBER",
          "STRING",
          "SELECT_NUMBER",
          "SELECT_STRING",
          "TAG",
          "ENTITY",
          "LINK"
        ]
      },
      "EnsureTagSetRequest": {
        "type": "object",
        "description": "Request to provision (get-or-create) the caller's tag set for a scope.",
        "required": ["scope"],
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/TagScope"
          }
        }
      },
...
      "EntityType": {
        "type": "string",
        "description": "Type of entity that can be referenced by entity properties.",
        "enum": [
          "CALENDAR_EVENT",
          "CALL_RECORD",
          "CHANNEL",
          "CHAT",
          "COMPANY",
          "DOCUMENT",
          "PROJECT",
          "TASK",
          "THREAD",
          "USER"
        ]
      },
...
      "MergeTagRequest": {
        "type": "object",
        "description": "Request to replace a personal label with an existing team label.",
        "required": ["option_id", "target_option_id"],
        "properties": {
          "option_id": {
            "type": "string",
            "format": "uuid",
            "description": "The personal label to retire."
          },
          "target_option_id": {
            "type": "string",
            "format": "uuid",
            "description": "The team label every entity carrying `option_id` is retagged with."
          }
        }
      },
      "PromoteTagRequest": {
        "type": "object",
        "description": "Request to share one of the caller's personal labels with their team.",
        "required": ["option_id"],
        "properties": {
          "option_id": {
            "type": "string",
            "format": "uuid",
            "description": "The personal label to move into the team tag set."
          }
        }
      },
...
      "PropertyDefinitionDetailResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PropertyOwner"
          },
          {
            "type": "object",
            "required": [
              "id",
              "displayName",
              "dataType",
              "isMultiSelect",
              "isSystem",
              "isMetadata"
            ],
            "properties": {
              "createdAt": {
                "type": ["string", "null"],
                "format": "date-time"
              },
              "dataType": {
                "$ref": "#/components/schemas/DataType"
              },
              "displayName": {
                "type": "string"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "isMetadata": {
                "type": "boolean",
                "description": "Flag to indicate if this is a system-generated metadata property"
              },
              "isMultiSelect": {
                "type": "boolean"
              },
              "isSystem": {
                "type": "boolean",
                "description": "Flag to indicate if this is a system property (managed by the system)"
              },
              "specificEntityType": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/EntityType"
                  }
                ]
              },
              "updatedAt": {
                "type": ["string", "null"],
                "format": "date-time"
              }
            }
          }
        ],
        "description": "Property definition response (API representation)."
      },
...
      "PropertyOptionResponse": {
        "type": "object",
        "description": "Property option response (API representation).",
        "required": ["id", "propertyDefinitionId", "displayOrder", "value"],
        "properties": {
          "color": {
            "type": ["string", "null"]
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "propertyDefinitionId": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "$ref": "#/components/schemas/PropertyOptionValue"
          }
        }
      },
      "PropertyOptionValue": {
        "oneOf": [
          {
            "type": "object",
            "description": "String value for SelectString properties",
            "required": ["value", "type"],
            "properties": {
              "type": {
                "type": "string",
                "enum": ["string"]
              },
              "value": {
                "type": "string",
                "description": "String value for SelectString properties"
              }
            }
          },
          {
            "type": "object",
            "description": "Number value for SelectNumber properties",
            "required": ["value", "type"],
            "properties": {
              "type": {
                "type": "string",
                "enum": ["number"]
              },
              "value": {
                "type": "number",
                "format": "double",
                "description": "Number value for SelectNumber properties"
              }
            }
          }
        ],
        "description": "The value of a property option - either a string or a number."
      },
      "PropertyOwner": {
        "oneOf": [
          {
            "type": "object",
            "description": "User-scoped property.",
            "required": ["user_id", "scope"],
            "properties": {
              "scope": {
                "type": "string",
                "enum": ["user"]
              },
              "user_id": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "Team-scoped property.",
            "required": ["team_id", "scope"],
            "properties": {
              "scope": {
                "type": "string",
                "enum": ["team"]
              },
              "team_id": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "type": "object",
            "description": "System-owned property (no user or team owner).",
            "required": ["scope"],
            "properties": {
              "scope": {
                "type": "string",
                "enum": ["system"]
              }
            }
          }
        ],
        "description": "Defines who owns a property - user-scoped, team-scoped, or system."
      },
...
      "TagPromotionConflictResponse": {
        "type": "object",
        "description": "Body returned when a promotion collides with a label the team already has.\nThe caller confirms with the user, then calls the merge endpoint with\n`conflicting_option.id` as the target.",
        "required": ["message", "conflicting_option"],
        "properties": {
          "conflicting_option": {
            "$ref": "#/components/schemas/PropertyOptionResponse",
            "description": "The team label that already uses this name."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the collision."
          }
        }
      },
      "TagScope": {
        "type": "string",
        "description": "Which owner a tag set belongs to.",
        "enum": ["user", "team"]
      },
      "TagSetResponse": {
        "type": "object",
        "description": "A tag set the caller can use. `definition` is absent until the set is provisioned\n(on first label create), in which case `options` is empty.",
        "required": ["scope", "options"],
        "properties": {
          "definition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PropertyDefinitionDetailResponse"
              }
            ]
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyOptionResponse"
            }
          },
          "scope": {
            "$ref": "#/components/schemas/TagScope"
          }
        }
      },
...
    }
  },
...
}